move shoplist code into shoplist page
